DAH-1139 move hmi table into own component
